Scanning: removed PatternSpace object (no need anymore for confidence, usage or hits maps), all the processing is done on PatternList objects.
CascadeMachine: remove the global threshold (needed by PatternSpace).

Scanning: updated all the objects to use PatternList and not PatternSpace.

Scanning: work on  MeanShiftSelector: not using LSH but a grid structure
to find nearest neighbours and correctly updating the points now (weights).

Scanning: it is possible now to use GreedyExplorer to generate a profile
around a specified sub-window.