Modified the library and the examples cmake files as to include JUST one header. Now the main CMakeLists.txt will create an /include directory where it will put all the headers and torch5spro.h file to include all these headers from /include.
This means ONLY torch5spro.h must be included now. And also now there is no need
to include all the spCore/ipCore in spCores.h and all the machines in Machines.h
as to be registered to the manager.

Moved Variable from the VariableCollector's header to .cc => a lot of loops!

Solve the bug for the segmentation fault in the scanning: @Seb - the Machines
should not deallocate the given spCores, because they are managed already.